pyproject.toml
[tool.poetry]
name = "upl-docs"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
django = "^5.1.5"
django-filter = "^24.3"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
pytest-cov = "^6.0.0"
requests = "^2.32.3"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
django-cors-headers = "^4.6.0"
drf-yasg = "^1.21.8"
redis = "^5.2.1"
flake8 = "^7.1.1"
isort = "^5.13.2"
black = "^24.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
